[UNTESTED] Add MTK GED KPI support to fix broken Mediatek gpufreq
Mediatek GPU scheduler likes to have the timestamps of the frames to be
able to adjust DVFS.
Technically this isn't completely needed, because it looks like DVFS can
work once we started triggering /proc/ged. But now that the work is
done, let's do it completely.

In benchmarks on Poco X4 GT, before this patch result is 1500, after is
5700. If we disable the patch after enabling it without rebooting, it
goes to 5400. So looks like GED KPI thingy does try to do a bit more
than just standard DVFS.

Thanks @sarthakroy2002 for the help and support (other people helped as
well)

[claxten10]:
 - Enable by default.

[dodyirawan85]:
 - Guard changes with soong config

Change-Id: Ic29a231ea8651efd598083611197aaa9e3c1fbbe
Signed-off-by: Saikrishna1504 <saikrishna26918@gmail.com>

Ported from AxionAOSP lineage-23.2 (9d2690a84), which puts the code
behind the libgui soong config variable support_mtk_ged_kpi. mt6878-devs
carried the same patch in their lineage-23.0 frameworks/native; it was
lost when malachite moved to 23.2.

Status: untested. Not built or run on a device.